feat(examples): merged basic setting example
