--- examples/agent-integration/forms/settings-form-rendering/pyproject.toml
@@ -1,39 +1,0 @@
-[project]
-name = "settings-form-rendering"
-version = "0.1.0"
-description = "Example demonstrarting agent settings functionality."
-authors = [
-    { name = "IBM Corp." },
-]
-requires-python = ">=3.13,<3.14"
-dependencies = [
-    "agentstack-sdk==0.5.3",
-    "pyyaml>=6.0.2",
-]
-
-[tool.ruff]
-line-length = 120
-target-version = "py313"
-
-[tool.uv.sources]
-agentstack-sdk = { path = "../../../../apps/agentstack-sdk-py", editable = true }
-
-[project.scripts]
-server = "settings_form_rendering.agent:run"
-
-[tool.pyrefly]
-project-includes = [
-    "**/*.py*",
-    "**/*.ipynb",
-]
-search-path = ['../../../../apps/agentstack-sdk-py/src', '.venv/lib/python3.13/site-packages']
-
-[build-system]
-requires = ["uv_build>=0.9.0,<0.10.0"]
-build-backend = "uv_build"
-
-[dependency-groups]
-dev = [
-    "pyrefly>=0.52.0",
-    "watchfiles>=1.1.0",
-]
